LPS-115426 Rename FormRenderer component in favor of the Pages and FormNoopRenderer to FormNoop
