+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/FormBuilder/FormBuilder.es.js
@@ -13,7 +13,7 @@
  */
 
 import {
-	FormNoopRenderer,
+	FormNoop,
 	PagesVisitor,
 	compose,
 	getConnectedReactComponentAdapter,
@@ -29,9 +29,7 @@ import withMoveableFields from './withMoveableFields.es';
 import withMultiplePages from './withMultiplePages.es';
 import withResizeableColumns from './withResizeableColumns.es';
 
-const FormNoopRendererAdapter = getConnectedReactComponentAdapter(
-	FormNoopRenderer
-);
+const FormNoopAdapter = getConnectedReactComponentAdapter(FormNoop);
 
 /**
  * Builder.
@@ -98,7 +96,7 @@ class FormBuilderBase extends Component {
 		return (
 			<div class="ddm-form-builder-wrapper">
 				<div class="container ddm-form-builder">
-					<FormNoopRendererAdapter
+					<FormNoopAdapter
 						activePage={activePage}
 						allowNestedFields={allowNestedFields}
 						dnd={dnd}
