+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/util/fields.es.js
@@ -13,10 +13,10 @@
  */
 
 import {
+	FormSupport,
 	PagesVisitor,
 	normalizeFieldName,
 } from 'dynamic-data-mapping-form-renderer';
-import {findFieldByFieldName} from 'dynamic-data-mapping-form-renderer/js/components/FormRenderer/FormSupport.es';
 
 export const generateFieldName = (
 	pages,
@@ -27,7 +27,7 @@ export const generateFieldName = (
 	let counter = 0;
 	let fieldName = normalizeFieldName(desiredName);
 
-	let existingField = findFieldByFieldName(pages, fieldName);
+	let existingField = FormSupport.findFieldByFieldName(pages, fieldName);
 
 	while (
 		(existingField && existingField.fieldName !== currentName) ||
@@ -37,7 +37,7 @@ export const generateFieldName = (
 			fieldName = normalizeFieldName(desiredName) + counter;
 		}
 
-		existingField = findFieldByFieldName(pages, fieldName);
+		existingField = FormSupport.findFieldByFieldName(pages, fieldName);
 
 		counter++;
 	}
