+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/util/FormSupport.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {PagesVisitor} from '../../util/visitors.es';
+import {PagesVisitor} from './visitors.es';
 
 export const implAddColumn = (size, fields = []) => {
 	return {
